+++ src/main/java/com/bsuir/buspark/bl/impl/CityServiceImpl.java
@@ -1,8 +1,10 @@
 package com.bsuir.buspark.bl.impl;
 
 import com.bsuir.buspark.bl.CityService;
-import com.bsuir.buspark.bl.exception.CityNotFoundException;
+import com.bsuir.buspark.bl.exception.notFound.CityNotFoundException;
+import com.bsuir.buspark.bl.exception.other.CityNowInUseException;
 import com.bsuir.buspark.dal.CityRepository;
+import com.bsuir.buspark.dal.TicketRepository;
 import com.bsuir.buspark.entity.City;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,8 @@ import java.util.List;
 public class CityServiceImpl implements CityService {
     @Autowired
     private CityRepository cityRepository;
+    @Autowired
+    private TicketRepository ticketRepository;
 
     @Override
     public City create(City city) {
@@ -38,10 +42,18 @@ public class CityServiceImpl implements CityService {
 
     @Override
     public void delete(int cityId) {
-        cityRepository.findById(cityId).orElseThrow(
+        City city = cityRepository.findById(cityId).orElseThrow(
                 () -> new CityNotFoundException("Cannot delete city with requested ID")
         );
-        cityRepository.deleteById(cityId);
+
+        if (ticketRepository.findByArrivalCity(city).isEmpty() && ticketRepository.findByDepartmentCity(city).isEmpty())
+        {
+            cityRepository.deleteById(cityId);
+        }
+        else
+        {
+            throw new CityNowInUseException("City now in use");
+        }
     }
 
     @Override
