+++ src/main/java/com/bsuir/buspark/bl/impl/RoleServiceImpl.java
@@ -1,8 +1,11 @@
 package com.bsuir.buspark.bl.impl;
 
 import com.bsuir.buspark.bl.RoleService;
-import com.bsuir.buspark.bl.exception.RoleNotFoundException;
+import com.bsuir.buspark.bl.UserService;
+import com.bsuir.buspark.bl.exception.notFound.RoleNotFoundException;
+import com.bsuir.buspark.bl.exception.other.RoleAlreadyExistsException;
 import com.bsuir.buspark.dal.RoleRepository;
+import com.bsuir.buspark.dal.UserRepository;
 import com.bsuir.buspark.entity.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +17,9 @@ public class RoleServiceImpl implements RoleService {
 
     @Autowired
     private RoleRepository roleRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     @Override
     public Role create(Role role) {
@@ -27,7 +33,6 @@ public class RoleServiceImpl implements RoleService {
                 () -> new RoleNotFoundException("Cannot update role with requested ID")
         );
         selectedRole.setName(role.getName());
-        selectedRole.setUsers(role.getUsers());
         return roleRepository.save(selectedRole);
     }
 
@@ -40,14 +45,27 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void delete(int roleId) {
-        roleRepository.findById(roleId).orElseThrow(
+        Role role = roleRepository.findById(roleId).orElseThrow(
                 () -> new RoleNotFoundException("Cannot delete role with requested ID")
         );
-        roleRepository.deleteById(roleId);
+
+        if (userRepository.findByRolesContains(role) == null)
+        {
+            roleRepository.deleteById(roleId);
+        }
+        else
+        {
+            throw new RoleAlreadyExistsException("Role is in use");
+        }
     }
 
     @Override
     public List<Role> getAll() {
         return roleRepository.findAll();
+    }
+
+    @Override
+    public List<Role> findByName(String name) {
+        return roleRepository.findByName(name);
     }
 }
